+++ assets/js/dashboard/stats/current-visitors.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom'
+import { countFilters } from '../query';
 
 export default class CurrentVisitors extends React.Component {
   constructor(props) {
@@ -22,10 +23,12 @@ export default class CurrentVisitors extends React.Component {
   }
 
   render() {
+    if (countFilters(this.props.query) !== 0) { return null }
+
     const { currentVisitors } = this.state;
     if (currentVisitors !== null) {
       return (
-        <Link to={`/${encodeURIComponent(this.props.site.domain)}?period=realtime`} className="block text-sm font-bold text-gray-500 dark:text-gray-300">
+        <Link to={`/${encodeURIComponent(this.props.site.domain)}?period=realtime`} className="block text-sm font-bold text-gray-500 dark:text-gray-300 mr-auto ml-2">
           <svg className="w-2 mr-2 fill-current text-green-500 inline" viewBox="0 0 16 16" xmlns="http://www.w3.org/2000/svg">
             <circle cx="8" cy="8" r="8"/>
           </svg>