+++ assets/tailwind.config.js
@@ -28,6 +28,12 @@ module.exports = {
       },
       zIndex: {
         '9': 9,
+<<<<<<< HEAD
+=======
+      },
+      maxWidth: {
+        '2xs': '16rem',
+>>>>>>> 9bbaa756b8241044818969eb5621468ef488721e
       }
     },
   },
@@ -39,6 +45,8 @@ module.exports = {
       borderWidth: ['dark'],
       backgroundOpacity: ['dark'],
       display: ['dark'],
+      cursor: ['hover'],
+      justifyContent: ['responsive']
     }
   },
   plugins: [